Added scaffolding for clear, renamed move.h
Added the cpp and header files for clear, and called it at the correct spot in main. Added it to cmakelists.
Also renamed move.h to matrix.h because all of its contents relate to the matrix. This means it makes more sense to include it in clear (including matrix rather than move, sounds better).
