+++ headers/clear.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "const.h"
+#include "matrix.h"
+void clearRows(matrixType &matrix);
