+++ src/clear.cpp
@@ -1,0 +1,4 @@
+#include "../headers/clear.h"
+#include "../headers/const.h"
+
+void clearRows(matrixType &matrix) { return; }
