+++ src/move.cpp
@@ -1,5 +1,6 @@
-#include "../headers/move.h"
 #include "../headers/const.h"
+#include "../headers/matrix.h"
+
 
 coords movePiece(matrixType &matrix, pieceCoords piece, char direction,
                  coords offset) {
